requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
scikit-learn>=1.2.0
Pillow>=9.0.0
joblib>=1.2.0
